fixed build toolchain related issues
